+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 Changelog
 =========
 
+* **2013-09-22**: [QueryBuilder] Renamed fieldExists => fieldIsset.
+* **2013-09-13**: [QueryBuilder] Replaced query builder with new
+  implementation. See new documentation.
 * **2013-08-16**: [Model] removed Doctrine\ODM\PHPCR\Document\Image and moved
   it to the CmfMediaBundle. Note that the CmfMediaBundle image is itself a file
   and no longer uses a file child.